# PERF: Timestamp dont create ndarray