```
In [1]: import pandas as pd                                                                                                                                   
In [2]: ts = pd.Timestamp.now()                                                                                                                               

In [3]: %timeit ts.day_name()                                                                                                                                 
6.46 µs ± 158 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)  # <-- PR
8.39 µs ± 107 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)  # <-- master
```